Re-word comparison to not be in the negative.
I think it might be easier to read considering all the other features are written in the positive and it took a second to try understand what the 'Yes' and 'No's meant in that case. I have re-worded it in a way that I think still keeps the same logic please do double check as I don't know for sure if I am right.